backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
deepface==0.0.79
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.4
mediapipe==0.10.7
Pillow==10.1.0
tf-keras==2.15.0
tensorflow==2.15.0
python-dotenv==1.0.0
